Made reusable Algorithm component
Not the prettiest thing by far, though. I wasn't able to pass in a JS object as a property or similar, which means the Algorithm component instead has to assume the current data context represents the algorithm, and will simply bind to fields on that and nothing else. Additionally, in order to not duplicate the file name transforms from the algorithms list, I introduced the Database concept that represents the "source of truth". Not sure I'm too happy with it, but it works and will do for now.